Moves the console to the bottom, removes the right terminal
