+++ lib/archethic_playground_web/live/components/sidebar_component.ex
@@ -54,7 +54,6 @@ defmodule ArchethicPlaygroundWeb.SidebarComponent do
         <a
                 class="inline-flex items-center w-full text-sm font-semibold transition-colors duration-150 hover:text-gray-800 dark:hover:text-gray-200"
                 href="#"
-                @click="$dispatch('open-terminal',{id})"
                 data-tippy-content="Validate Contract"
                 id="id_ValidateContract"
                 phx-hook="hook_ValidateContract"
